pyproject.toml
[project]
name = "jukebox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "python-telegram-bot>=22.5",
    "yt-dlp>=2025.12.8",
    "ytmusicapi>=1.11.4",
]
